feat(vtex): split timeline and sales card into separate MCP App tools
VTEX_ORDERS_TIMELINE renders only the hourly chart; VTEX_ORDERS_SALES_CARD
accepts period today/last_1h/last_5min for composable home widgets. Shared
OMS _stats=1 logic lives in orders-oms.ts.
