+++ vtex/README.md
@@ -41,7 +41,8 @@ These handle multi-step operations not available in the generated SDK:
 | `VTEX_SEARCH_COLLECTIONS` | Search collections by name or terms |
 | `VTEX_REORDER_COLLECTION` | Reorder collections with SKU/product IDs via XML import |
 | `VTEX_UPDATE_PRODUCT_SPECIFICATIONS` | Bulk replace product specifications |
-| `VTEX_ORDERS_TIMELINE` | Today's orders by hour with sales summary cards (MCP App UI) |
+| `VTEX_ORDERS_TIMELINE` | Hourly orders bar chart for today (MCP App UI) |
+| `VTEX_ORDERS_SALES_CARD` | Sales card for `today`, `last_1h`, or `last_5min` (MCP App UI) |
 
 ## Authentication
