+++ vtex/server/tools/index.ts
@@ -82,6 +82,7 @@ import {
 
 // ── Custom tools ──────────────────────────────────────────────────────────────
 import { ordersTimeline } from "./custom/orders-timeline.ts";
+import { ordersSalesCard } from "./custom/orders-sales-card.ts";
 import { searchCollections } from "./custom/search-collections.ts";
 import { reorderCollection } from "./custom/reorder-collection.ts";
 import { updateProductSpecifications } from "./custom/update-product-specifications.ts";
@@ -163,8 +164,10 @@ const registryFactories = [
 ];
 
 const customFactories = [
-  // Today's orders timeline with MCP App UI
+  // Today's orders timeline chart (MCP App UI)
   ordersTimeline,
+  // Sales summary card for today / last hour / last 5 min (MCP App UI)
+  ordersSalesCard,
   // Collection search (endpoint absent from generated SDKs)
   searchCollections,
   // Collection overwrite/reorder via XML import flow
